refactor: move constants to another file
